refactor(#741): refactor subjectController
